Wire up top-level template files at note creation time
Given a template dir, the plugin is now configurable
for multiple prefix matchers. For any top-level
file in the given templates directory setting,
it will be selectable via the dropdown in each
prefix matcher config line.

Whenever a new file is created that starts with
a prefix match in the settings, it will be automatically
populated with the corresponding template.